Split out Translation & TranslationQuerySet
